Extract FederatedSchemaTrait
